Create C program to compute the profit given X, A and B.
Each Sunday, a newspaper agency sells X copies of a certain newspaper for Rs.A per copy. The cost to the agency of each newspaper is Rs.B.

The agency pays a fixed cost for storage, delivery and soon of Rs.100 per Sunday. The newspaper agency wants to calculate the profit obtained on Sundays.

Can you please help them by writing a C program to compute the profit given X, A and B.

Sample Input and Output:

Number of copies sold : 1000
Cost of each copy : 2
Cost spent by agency on each newspaper : 1
The profit obtained : Rs.900